+++ src/main/java/org/aion4j/avm/idea/action/local/LocalGetBalanceAction.java
@@ -4,8 +4,6 @@ import com.intellij.notification.NotificationType;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.components.ServiceManager;
 import com.intellij.openapi.project.Project;
-import org.aion4j.avm.idea.action.AvmBaseAction;
-import org.aion4j.avm.idea.action.remote.GetBalanceAction;
 import org.aion4j.avm.idea.misc.AvmIcons;
 import org.aion4j.avm.idea.misc.IdeaUtil;
 import org.jetbrains.annotations.NotNull;
@@ -15,9 +13,7 @@ import org.jetbrains.idea.maven.execution.MavenRunnerSettings;
 
 import javax.swing.*;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 public class LocalGetBalanceAction extends AvmLocalBaseAction {
 
@@ -36,7 +32,7 @@ public class LocalGetBalanceAction extends AvmLocalBaseAction {
         List<String> goals = new ArrayList<>();
         goals.add("aion4j:get-balance");
 
-        MavenRunnerParameters mavenRunnerParameters = getMavenRunnerParameters(project, goals);
+        MavenRunnerParameters mavenRunnerParameters = getMavenRunnerParameters(e, project, goals);
         MavenRunnerSettings mavenRunnerSettings = getMavenRunnerSettings(project);
 
         mavenRunner.run(mavenRunnerParameters, mavenRunnerSettings, () -> {