+++ src/main/java/org/aion4j/avm/idea/action/local/LocalGetBalanceByAccountAction.java
@@ -5,7 +5,6 @@ import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.components.ServiceManager;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.text.StringUtil;
-import org.aion4j.avm.idea.action.AvmBaseAction;
 import org.aion4j.avm.idea.action.local.ui.LocalGetAccountDialog;
 import org.aion4j.avm.idea.misc.AvmIcons;
 import org.aion4j.avm.idea.misc.IdeaUtil;
@@ -16,9 +15,7 @@ import org.jetbrains.idea.maven.execution.MavenRunnerSettings;
 
 import javax.swing.*;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 public class LocalGetBalanceByAccountAction extends AvmLocalBaseAction {
 
@@ -52,7 +49,7 @@ public class LocalGetBalanceByAccountAction extends AvmLocalBaseAction {
         List<String> goals = new ArrayList<>();
         goals.add("aion4j:get-balance");
 
-        MavenRunnerParameters mavenRunnerParameters = getMavenRunnerParameters(project, goals);
+        MavenRunnerParameters mavenRunnerParameters = getMavenRunnerParameters(e, project, goals);
         MavenRunnerSettings mavenRunnerSettings = getMavenRunnerSettings(project);
 
         mavenRunnerSettings.getMavenProperties().put("address", account);